refactor(profile): move changePasswordSchema from auth to profile feature
- changePassword.schema.ts: schema moved to profile/schemas
- changePassword.schema.spec.ts: tests moved to profile/schemas
- index.ts: barrel export for profile schemas
- ChangePasswordForm.tsx: import from own feature instead of auth
- auth.schema.ts: remove changePassword schema and type
- auth.schema.spec.ts: remove changePassword tests
- auth/index.ts: remove changePassword exports

Closes #7
