+++ apps/web/src/features/auth/schemas/auth.schema.spec.ts
@@ -1,12 +1,6 @@
 import { describe, expect, it } from 'bun:test'
 
-import {
-  changePasswordSchema,
-  forgotPasswordSchema,
-  loginSchema,
-  resetPasswordSchema,
-  signupSchema
-} from './auth.schema'
+import { forgotPasswordSchema, loginSchema, resetPasswordSchema, signupSchema } from './auth.schema'
 
 describe('loginSchema', () => {
   it('should validate a valid login request', () => {
@@ -216,35 +210,6 @@ describe('resetPasswordSchema', () => {
     const result = resetPasswordSchema.safeParse({
       password: 'short',
       confirmPassword: 'short'
-    })
-    expect(result.success).toBe(false)
-  })
-})
-
-describe('changePasswordSchema', () => {
-  it('should validate valid change password request', () => {
-    const result = changePasswordSchema.safeParse({
-      currentPassword: 'oldpass123',
-      newPassword: 'newpass123',
-      confirmPassword: 'newpass123'
-    })
-    expect(result.success).toBe(true)
-  })
-
-  it('should reject when new passwords do not match', () => {
-    const result = changePasswordSchema.safeParse({
-      currentPassword: 'oldpass123',
-      newPassword: 'newpass123',
-      confirmPassword: 'different123'
-    })
-    expect(result.success).toBe(false)
-  })
-
-  it('should reject empty current password', () => {
-    const result = changePasswordSchema.safeParse({
-      currentPassword: '',
-      newPassword: 'newpass123',
-      confirmPassword: 'newpass123'
     })
     expect(result.success).toBe(false)
   })
