+++ apps/web/src/features/auth/schemas/auth.schema.ts
@@ -44,24 +44,9 @@ export const resetPasswordSchema = z
     path: ['confirmPassword']
   })
 
-export const changePasswordSchema = z
-  .object({
-    currentPassword: z.string().min(1, { message: 'Current password is required' }),
-    newPassword: z
-      .string()
-      .min(1, { message: 'New password is required' })
-      .min(8, { message: 'Password must be at least 8 characters' }),
-    confirmPassword: z.string().min(1, { message: 'Please confirm your password' })
-  })
-  .refine(data => data.newPassword === data.confirmPassword, {
-    message: 'Passwords do not match',
-    path: ['confirmPassword']
-  })
-
 // Types
 export type LoginSchema = z.infer<typeof loginSchema>
 export type SignupSchema = z.infer<typeof signupSchema>
 export type AuthSchema = LoginSchema | SignupSchema
 export type ForgotPasswordSchema = z.infer<typeof forgotPasswordSchema>
 export type ResetPasswordSchema = z.infer<typeof resetPasswordSchema>
-export type ChangePasswordSchema = z.infer<typeof changePasswordSchema>
