+++ apps/web/src/features/profile/components/ChangePasswordForm.tsx
@@ -3,9 +3,10 @@ import { useState } from 'react'
 import { useForm } from 'react-hook-form'
 
 import { Button, FormWrapper, Input } from '~/components/ui'
-import type { ChangePasswordSchema } from '~/features/auth'
-import { changePasswordSchema } from '~/features/auth'
 import { authClient } from '~/lib/authClient'
+
+import type { ChangePasswordSchema } from '../schemas'
+import { changePasswordSchema } from '../schemas'
 
 export const ChangePasswordForm = () => {
   const [status, setStatus] = useState<'idle' | 'success' | 'error'>('idle')
