+++ apps/web/src/features/profile/schemas/changePassword.schema.ts
@@ -1,0 +1,17 @@
+import { z } from 'zod'
+
+export const changePasswordSchema = z
+  .object({
+    currentPassword: z.string().min(1, { message: 'Current password is required' }),
+    newPassword: z
+      .string()
+      .min(1, { message: 'New password is required' })
+      .min(8, { message: 'Password must be at least 8 characters' }),
+    confirmPassword: z.string().min(1, { message: 'Please confirm your password' })
+  })
+  .refine(data => data.newPassword === data.confirmPassword, {
+    message: 'Passwords do not match',
+    path: ['confirmPassword']
+  })
+
+export type ChangePasswordSchema = z.infer<typeof changePasswordSchema>
