+++ apps/web/src/features/profile/schemas/index.ts
@@ -1,0 +1,2 @@
+export { changePasswordSchema, type ChangePasswordSchema } from './changePassword.schema'
+export { updateProfileSchema, type UpdateProfileSchema } from './user.schema'
